+++ IDotMatrixRenderer.h
@@ -60,7 +60,7 @@ public:
     uint8_t glyphCount,
     uint8_t glyphWidth,
     uint8_t glyphHeight,
-    uint8_t glyphBytes,
+    uint16_t glyphBytes,
     uint8_t motionEffect,
     uint8_t speed,
     uint8_t colorMode,
@@ -135,7 +135,7 @@ private:
   uint8_t textGlyphCount_ = 0;
   uint8_t textGlyphWidth_ = 0;
   uint8_t textGlyphHeight_ = 0;
-  uint8_t textGlyphBytes_ = 0;
+  uint16_t textGlyphBytes_ = 0;
   uint8_t textMotionEffect_ = 0;
   uint8_t textSpeed_ = 5;
   uint8_t textColorMode_ = 1;
